+++ src/util.h
@@ -51,7 +51,7 @@ inline void push_back(std::vector<uint8_t>& values, uint16_t value)
 constexpr int32_t log_2(int32_t n) noexcept
 {
     int32_t x = 0;
-    while (n > (static_cast<int32_t>(1) << x))
+    while (n > (1 << x))
     {
         ++x;
     }
